requirements.txt
flask
ccxt
pandas
psycopg2-binary
websocket-client
requests
gunicorn
Flask-Cors
